requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.9
gspread>=5.10.0
oauth2client>=4.1.3
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
